refactor: move items concerno to ListTable
